Fix ME in PushAndCalConcurrency.
`PushAndCalConcurrency` is called in a loop.
And it assumes that it has been called under lock-protection.
However, since `PushAndCalConcurrency` itself can unlock the lock
in some execution path, the subsequent call of `PushAndCalConcurrency`
will not happen under the lock protection causing mutual exclusion violation on
`whenMapVec_` when other threads call `GetHeadTask`.

The current solution is to not unlock if `PushAndCalConcurrency` is to be
called in a loop. Using the param flag `needUnlock`.

Issue: https://gitee.com/openharmony/resourceschedule_ffrt/issues/ICOXHF

Signed-off-by: Lilith Oberhauser <lilith.oberhauser@huawei.com>